+++ test/dummy/test/integration/playground_test.rb
@@ -77,8 +77,9 @@ class PlaygroundTest < ActionDispatch::IntegrationTest
     assert_select "[data-flat-pack--collapse-target='trigger']", text: /Find page/
     assert_select "[data-flat-pack--collapse-target='trigger']", text: /Done/
     assert_select "[data-flat-pack--collapse-target='trigger']", text: /Reply/
+    assert_select "[id^='playground-step-'][id$='-content']", text: /Find the Getting Started page\./
+    assert_select "[id^='playground-step-'][id$='-content']", text: /Finished\./
     assert_select "#playground-reply-content[hidden]", text: /Finished\./
-    assert_select "#playground-reply-content", text: /Used Find page\./, count: 0
     assert_select "h2", text: "Page librarian", count: 0
     assert_select "h2", text: "Skills", count: 0
     assert_select "h2", text: "Reply", count: 0
@@ -139,8 +140,8 @@ class PlaygroundTest < ActionDispatch::IntegrationTest
     get path
 
     assert_select "[data-flat-pack--collapse-target='trigger']", text: /Find page/
+    assert_select "[id^='playground-step-'][id$='-content']", text: /Find the Getting Started page\./
     assert_select "#playground-reply-content[hidden]", text: /Finished\./
-    assert_select "[id^='playground-step-'][id$='-content']", text: /Used Find page\./
     run = RecordingStudioAgents::AgentRun.find_by!(
       root_recording_id: @root.id,
       idempotency_key: path.split("/").last